+++ src/mkvtoolnix-gui/chapter_editor/tool.cpp
@@ -193,7 +193,7 @@ Tool::selectFileToOpen(bool append) {
 #endif  // HAVE_DVDREAD
 
   auto fileNames = Util::getOpenFileNames(this, append ? QY("Append files in chapter editor") : QY("Open files in chapter editor"), Util::Settings::get().lastOpenDirPath(),
-                                          QY("Supported file types")           + Q(" (*.cue%1 *.mpls *.mkv *.mka *.mks *.mk3d *.txt *.webm *.xml);;").arg(ifo) +
+                                          QY("Supported file types")           + Q(" (*.cue%1 *.meta *.mpls *.mkv *.mka *.mks *.mk3d *.txt *.webm *.xml);;").arg(ifo) +
                                           QY("Matroska files")                 + Q(" (*.mkv *.mka *.mks *.mk3d);;") +
                                           QY("WebM files")                     + Q(" (*.webm);;") +
                                           QY("Blu-ray playlist files")         + Q(" (*.mpls);;") +
@@ -201,6 +201,7 @@ Tool::selectFileToOpen(bool append) {
                                           QY("XML chapter files")              + Q(" (*.xml);;") +
                                           QY("Simple OGM-style chapter files") + Q(" (*.txt);;") +
                                           QY("Cue sheet files")                + Q(" (*.cue);;") +
+                                          QY("ffmpeg metadata files")          + Q(" (*.meta);;") +
                                           QY("All files")                      + Q(" (*)"));
   if (fileNames.isEmpty())
     return;